Discard output from ended process
Interestingly, it still seems to be possible to get a situation
where output is received from a process that is no longer running,
even after the last round of changes.

So far I have only observed this when the process has been explicitly
killed, so let’s try ignoring such data – but logging it – and see if
it ever happens in other circumstances.